@@ -9,17 +9,24 @@
   "license": "BSD-3-Clause",
   "type": "library",
   "require": {
-    "php": ">=7.3"
+    "php": ">=7.4"
   },
   "require-dev": {
-    "alex-kalanis/kw_mapper": ">=0.7 <3",
+    "alex-kalanis/kw_mapper": ">=4.0 <5",
     "dibi/dibi": "~4.2",
-    "doctrine/dbal": "~3.1",
+    "doctrine/dbal": "~3.8",
     "nette/database": "~3.1",
+    "shipmonk/composer-dependency-analyser": "^1.4",
     "phpstan/phpstan": "^1.0",
     "phpstan/phpstan-phpunit": "^1.0",
     "friendsofphp/php-cs-fixer": "^3.0",
-    "phpunit/phpunit": ">=7.2 <=9"
+    "phpunit/phpunit": ">=8.0 <=9"
+  },
+  "suggest": {
+    "alex-kalanis/kw_mapper": "Mapper for accessing resources in KWCMS",
+    "dibi/dibi": "Dibi - smart database abstraction layer",
+    "doctrine/dbal": "Doctrine Database Abstraction Layer",
+    "nette/database": "A database layer with a familiar PDO-like API but much more powerful",
   },
   "autoload": {
     "psr-4": {
