@@ -12,7 +12,7 @@ class FromWith extends AType
     public function setFiltering(string $colName, $value)
     {
         if ('' !== $value) {
-            $this->dibiFluent->where($colName . ' >= ?', $value);
+            $this->getSource()->where($colName . ' >= ?', $value);
         }
         return $this;
     }
