@@ -12,9 +12,9 @@ class To extends AType
     public function setFiltering(string $colName, $value)
     {
         if ('' !== $value) {
-            $this->queryBuilder->where($this->queryBuilder->expr()->lt(
+            $this->getSource()->where($this->getSource()->expr()->lt(
                 $colName,
-                $this->queryBuilder->createNamedParameter($value)
+                $this->getSource()->createNamedParameter($value)
             ));
         }
         return $this;
