@@ -12,7 +12,7 @@ class To extends AType
     public function setFiltering(string $colName, $value)
     {
         if ('' !== $value) {
-            $this->search->to($colName, $value, false);
+            $this->getSource()->to($colName, $value, false);
         }
         return $this;
     }
